+++ VCore/Extensions/Foundation/Array.CompactMapNonEmpty.swift
@@ -1,5 +1,5 @@
 //
-//  CompactMapNonEmpty.swift
+//  Array.CompactMapNonEmpty.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.