+++ VCore/Extensions/Foundation/Date.Component.swift
@@ -1,5 +1,5 @@
 //
-//  DateComponents.swift
+//  Date.Component.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.