+++ VCore/Extensions/Foundation/NSObject.NSObjectName.swift
@@ -1,5 +1,5 @@
 //
-//  NSObjectName.swift
+//  NSObject.NSObjectName.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.