+++ VCore/Extensions/Foundation/Sequence.KeyPathSort.swift
@@ -1,5 +1,5 @@
 //
-//  KeyPathSort.swift
+//  Sequence.KeyPathSort.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/13/21.
@@ -9,7 +9,7 @@ import Foundation
 
 // MARK: - KeyPath Sort
 extension Sequence {
-    /// Sorts sequence by a key path.
+    /// Returns the elements of the sequence, sorted using the given keypath as the comparison between elements.
     public func sorted<T: Comparable>(
         by keypath: KeyPath<Element, T>
     ) -> [Element] {