+++ VCore/Extensions/Foundation/StringProtocol.Subscript.swift
@@ -1,5 +1,5 @@
 //
-//  StringProtocolSubscripts.swift
+//  StringProtocol.Subscript.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.