+++ VCore/Extensions/SwiftUI/View.ConditionalModifiers.swift
@@ -1,5 +1,5 @@
 //
-//  ConditionalViewModifiers.swift
+//  View.ConditionalModifiers.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 10/6/21.