+++ VCore/Extensions/SwiftUI/View.CornerRadius.swift
@@ -1,5 +1,5 @@
 //
-//  CornerRadius.swift
+//  View.CornerRadius.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 10/6/21.