+++ VCore/Extensions/SwiftUI/View.CustomFrames.swift
@@ -1,5 +1,5 @@
 //
-//  CustomViewFrames.swift
+//  View.CustomFrames.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 10/6/21.