+++ VCore/Extensions/SwiftUI/View.StandardNavigationTitle.swift
@@ -1,5 +1,5 @@
 //
-//  StandardNavigationTitle.swift
+//  View.StandardNavigationTitle.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/21/21.
@@ -10,7 +10,7 @@ import SwiftUI
 // MARK: - Extension
 @available(iOS 14, *)
 extension View {
-    /// Configures the view’s title for purposes of navigation, using a string in a inline display mode
+    /// Configures the view’s title for purposes of navigation, using a string in a inline display mode.
     public func standardNavigationTitle(_ title: String) -> some View {
         self
             .modifier(StandardNavigationTitle(title))
@@ -27,11 +27,8 @@ fileprivate struct StandardNavigationTitle: ViewModifier {
     fileprivate init(_ title: String) {
         self.title = title
     }
-}
 
-// MARK: - Body
-@available(iOS 14, *)
-extension StandardNavigationTitle {
+    // MARK: Body
     fileprivate func body(content: Content) -> some View {
         content
             .navigationTitle(title)