+++ VCore/Extensions/UIKit/UILabel.Configuration.swift
@@ -1,5 +1,5 @@
 //
-//  LabelConfiguration.swift
+//  UILabel.Configuration.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/13/21.