+++ VCore/Extensions/UIKit/UINavigationBar.Height.swift
@@ -1,5 +1,5 @@
 //
-//  NavigationBarHeight.swift
+//  UINavigationBar.Height.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 9/12/21.