+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 #### [1.2.0(6)]
 
-- Project now contains additional files under `Extra`
+- Framework now contains additional files under `Extra`
 - `LocalizationService` has been added to `Extra`
 - XCode Templates have been added to `Extra`
+- `UILabel`'s `singleLineNaturalHeightConstant` has been renamed to `singleLineHeight`
+- `UITableView`'s `removeExtraSeparators` and `removeExtraAndLastSeparators` have been deprecated, as they have no effect
+- `UIScreen.rootView` has been deprecated in favor of `UIApplication.AppRootWindow` file under `Extra`
+- `UIScreen.safeAreaInsets` has been deprecated
 
 #### [1.1.0(5)](https://github.com/VakhoKontridze/VCore/releases/download/1.1.0/VCore.xcframework.zip) — *2021 10 30*